+++ Assets/Scripts/Demos/PerlinNoisePreview.cs
@@ -38,7 +38,7 @@ public class PerlinNoisePreview : GenericBehaviour
         Renderer renderer = perlinObject.GetComponent<MeshRenderer>();
         Texture2D texture = new Texture2D(width, height);
 
-        double[,] preColors = PGApi.Perlin.Generate2DMapUnity(width, height, scale, offsetx, offsety);
+        float[,] preColors = PGApi.Perlin.Generate2DMapUnity(width, height, scale, offsetx, offsety);
 
         for (int x = 0; x < width; x++)
             for (int y = 0; y < height; y++)
