+++ PG_Library/KenPerlin.cpp
@@ -8,7 +8,7 @@
 
 
 
-double ClassicPerlinNoise1D(double x)
+float ClassicPerlinNoise1D(float x)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -24,19 +24,19 @@ double ClassicPerlinNoise1D(double x)
 	int gi1 = g_HASH_TABLE_KEN_PERLIN[xi + 1] % 12;
 
 	// Calculate noise contributions from each of the eight corners
-	double n0 = Dot(g_GRAD3[gi0][0], x);
-	double n1 = Dot(g_GRAD3[gi1][0], x - 1);
+	float n0 = Dot(g_GRAD3[gi0][0], x);
+	float n1 = Dot(g_GRAD3[gi1][0], x - 1);
 
 	// Compute the fade curve value for each of x, y, z
-	double u = Fade(x);
+	float u = Fade(x);
 
 	// Interpolate along x the contributions from each of the corners
-	double nx = Lerpd(n0, n1, u);
+	float nx = Lerpd(n0, n1, u);
 
 	return nx;
 }
 
-double ClassicPerlinNoise2D(double x, double y)
+float ClassicPerlinNoise2D(float x, float y)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -57,26 +57,26 @@ double ClassicPerlinNoise2D(double x, double y)
 	int gi11 = g_HASH_TABLE_KEN_PERLIN[xi + 1 + g_HASH_TABLE_KEN_PERLIN[yi+1]] % 12;
 
 	// Calculate noise contributions from each of the eight corners
-	double n00 = DotN(2,g_GRAD3[gi00], x, y);
-	double n01 = DotN(2,g_GRAD3[gi01], x, y - 1);
-	double n10 = DotN(2,g_GRAD3[gi10], x - 1, y);
-	double n11 = DotN(2,g_GRAD3[gi11], x - 1, y - 1);
+	float n00 = DotN(2,g_GRAD3[gi00], x, y);
+	float n01 = DotN(2,g_GRAD3[gi01], x, y - 1);
+	float n10 = DotN(2,g_GRAD3[gi10], x - 1, y);
+	float n11 = DotN(2,g_GRAD3[gi11], x - 1, y - 1);
 
 	// Compute the fade curve value for each of x, y, z
-	double u = Fade(x);
-	double v = Fade(y);
+	float u = Fade(x);
+	float v = Fade(y);
 
 	// Interpolate along x the contributions from each of the corners
-	double nx0 = Lerpd(n00, n10, u);
-	double nx1 = Lerpd(n01, n11, u);
+	float nx0 = Lerpd(n00, n10, u);
+	float nx1 = Lerpd(n01, n11, u);
 
 	// Interpolate the four results along y
-	double nxy = Lerpd(nx0, nx1, v);
+	float nxy = Lerpd(nx0, nx1, v);
 
 	return nxy;
 }
 
-double ClassicPerlinNoise3D(double x, double y, double z)
+float ClassicPerlinNoise3D(float x, float y, float z)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -104,39 +104,39 @@ double ClassicPerlinNoise3D(double x, double y, double z)
 	int gi111 = g_HASH_TABLE_KEN_PERLIN[xi + 1 + g_HASH_TABLE_KEN_PERLIN[yi + 1 + g_HASH_TABLE_KEN_PERLIN[zi + 1]]] % 12;
 
 	// Calculate noise contributions from each of the eight corners
-	double n000 = DotN(3,g_GRAD3[gi000], x, y, z);
-	double n001 = DotN(3,g_GRAD3[gi001], x, y, z - 1);
-	double n010 = DotN(3,g_GRAD3[gi010], x, y - 1, z);
-	double n011 = DotN(3,g_GRAD3[gi011], x, y - 1, z - 1);
-	double n100 = DotN(3,g_GRAD3[gi100], x - 1, y, z);
-	double n101 = DotN(3,g_GRAD3[gi101], x - 1, y, z - 1);
-	double n110 = DotN(3,g_GRAD3[gi110], x - 1, y - 1, z);
-	double n111 = DotN(3,g_GRAD3[gi111], x - 1, y - 1, z - 1);
+	float n000 = DotN(3,g_GRAD3[gi000], x, y, z);
+	float n001 = DotN(3,g_GRAD3[gi001], x, y, z - 1);
+	float n010 = DotN(3,g_GRAD3[gi010], x, y - 1, z);
+	float n011 = DotN(3,g_GRAD3[gi011], x, y - 1, z - 1);
+	float n100 = DotN(3,g_GRAD3[gi100], x - 1, y, z);
+	float n101 = DotN(3,g_GRAD3[gi101], x - 1, y, z - 1);
+	float n110 = DotN(3,g_GRAD3[gi110], x - 1, y - 1, z);
+	float n111 = DotN(3,g_GRAD3[gi111], x - 1, y - 1, z - 1);
 
 	// Compute the fade curve value for each of x, y, z
-	double u = Fade(x);
-	double v = Fade(y);
-	double w = Fade(z);
+	float u = Fade(x);
+	float v = Fade(y);
+	float w = Fade(z);
 
 	// Interpolate along x the contributions from each of the corners
-	double nx00 = Lerpd(n000, n100, u);
-	double nx01 = Lerpd(n001, n101, u);
-	double nx10 = Lerpd(n010, n110, u);
-	double nx11 = Lerpd(n011, n111, u);
+	float nx00 = Lerpd(n000, n100, u);
+	float nx01 = Lerpd(n001, n101, u);
+	float nx10 = Lerpd(n010, n110, u);
+	float nx11 = Lerpd(n011, n111, u);
 
 	// Interpolate the four results along y
-	double nxy0 = Lerpd(nx00, nx10, v);
-	double nxy1 = Lerpd(nx01, nx11, v);
+	float nxy0 = Lerpd(nx00, nx10, v);
+	float nxy1 = Lerpd(nx01, nx11, v);
 
 	// Interpolate the two last results along z
-	double nxyz = Lerpd(nxy0, nxy1, w);
+	float nxyz = Lerpd(nxy0, nxy1, w);
 
 	return nxyz;
 }
 
 
 
-double PerlinNoiseND(int nDim, ...)
+float PerlinNoiseND(int nDim, ...)
 {
 	if (nDim < 1)
 	{
@@ -144,19 +144,19 @@ double PerlinNoiseND(int nDim, ...)
 	}
 
 	// Arrays
-	double* input = new double[nDim];                                       // Get the input from the function with unknown number of parameter
+	float* input = new float[nDim];                                       // Get the input from the function with unknown number of parameter
 	
 	// Get the input from the function with unknown number of parameters
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 
-	double result = PerlinNoiseNDArray(nDim, input);
+	float result = PerlinNoiseNDArray(nDim, input);
 
 	delete[] input;
 
 	return result;
 }
 
-double PerlinNoiseNDArray(int nDim, double* input)
+float PerlinNoiseNDArray(int nDim, float* input)
 {
 	if (nDim < 1)
 	{
@@ -174,7 +174,7 @@ double PerlinNoiseNDArray(int nDim, double* input)
 	// Arrays
 	int* unitGridCells = new int[nDim];                                     // Find unit grid cell containing point
 	int* gradientIndices = new int[NUMBER_OF_GRADIENT_INDICES];             // Calculate a set of eight hashed gradient indices
-	double* noiseContributions = new double[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
+	float* noiseContributions = new float[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
 	
 	// Find unit grid cell containing point
 	for (int i = 0; i < nDim; i++)
@@ -253,7 +253,7 @@ double PerlinNoiseNDArray(int nDim, double* input)
 	}
 
 	// Compute the fade curve value for each of x, y, z
-	double* curves = input;
+	float* curves = input;
 	for (int i = 0; i < nDim; i++)
 	{
 		curves[i] = Fade(input[i]);
@@ -269,7 +269,7 @@ double PerlinNoiseNDArray(int nDim, double* input)
 		}
 		interpolations /= 2;
 	}
-	double result = noiseContributions[0];
+	float result = noiseContributions[0];
 
 	delete[] unitGridCells;
 	delete[] gradientIndices;
@@ -279,14 +279,14 @@ double PerlinNoiseNDArray(int nDim, double* input)
 }
 
 
-double* ClassicPerlinNoise1D_Test_Input(double x)
+float* ClassicPerlinNoise1D_Test_Input(float x)
 {
-	double* result = new double[1];
+	float* result = new float[1];
 	result[0] = x;
 	return result;
 }
 
-double* PerlinNoiseND_Test_Input(int nDim, ...)
+float* PerlinNoiseND_Test_Input(int nDim, ...)
 {
 	if (nDim < 1)
 	{
@@ -294,14 +294,14 @@ double* PerlinNoiseND_Test_Input(int nDim, ...)
 	}
 
 	// Arrays
-	double* input = new double[nDim];                                       // Get the input from the function with unknown number of parameter
+	float* input = new float[nDim];                                       // Get the input from the function with unknown number of parameter
 
 	// Get the input from the function with unknown number of parameters
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 	return input;
 }
 
-int* ClassicPerlinNoise1D_Test_UnitGrid(double x)
+int* ClassicPerlinNoise1D_Test_UnitGrid(float x)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -319,10 +319,10 @@ int* PerlinNoiseND_Test_UnitGrid(int nDim, ...)
 	}
 
 	// Arrays
-	double* input = new double[nDim];                                       // Get the input from the function with unknown number of parameter
+	float* input = new float[nDim];                                       // Get the input from the function with unknown number of parameter
 
 	// Get the input from the function with unknown number of parameters
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 	// Constants that we will use along the function
 	const int N_DIM_MINUS_ONE = nDim - 1;
 	const int N_DIM_PLUS_ONE = nDim + 1;
@@ -343,7 +343,7 @@ int* PerlinNoiseND_Test_UnitGrid(int nDim, ...)
 }
 
 
-double* ClassicPerlinNoise1D_Test_Relative(double x)
+float* ClassicPerlinNoise1D_Test_Relative(float x)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -351,12 +351,12 @@ double* ClassicPerlinNoise1D_Test_Relative(double x)
 	// Get relative xyz coordinates of point within that cell
 	x = x - xi;
 
-	double* result = new double[1];
+	float* result = new float[1];
 	result[0] = x;
 	return result;
 }
 
-double* PerlinNoiseND_Test_Relative(int nDim, ...)
+float* PerlinNoiseND_Test_Relative(int nDim, ...)
 {
 	if (nDim < 1)
 	{
@@ -364,10 +364,10 @@ double* PerlinNoiseND_Test_Relative(int nDim, ...)
 	}
 
 	// Arrays
-	double* input = new double[nDim];                                       // Get the input from the function with unknown number of parameter
+	float* input = new float[nDim];                                       // Get the input from the function with unknown number of parameter
 
 	// Get the input from the function with unknown number of parameters
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 	// Constants that we will use along the function
 	const int N_DIM_MINUS_ONE = nDim - 1;
 	const int N_DIM_PLUS_ONE = nDim + 1;
@@ -380,7 +380,7 @@ double* PerlinNoiseND_Test_Relative(int nDim, ...)
 	int* unitGridCells = new int[nDim];                                     // Find unit grid cell containing point
 	int* gradientIndices = new int[NUMBER_OF_GRADIENT_INDICES];             // Calculate a set of eight hashed gradient indices
 	int* grad3 = new int[NUMBER_OF_EDGES * nDim];                           // Calculate the gradient helper for noise contributions
-	double* noiseContributions = new double[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
+	float* noiseContributions = new float[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
 
 	for (int i = 0; i < nDim; i++)
 	{
@@ -399,7 +399,7 @@ double* PerlinNoiseND_Test_Relative(int nDim, ...)
 }
 
 
-int* ClassicPerlinNoise1D_Test_Wrap(double x)
+int* ClassicPerlinNoise1D_Test_Wrap(float x)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -423,10 +423,10 @@ int* PerlinNoiseND_Test_Wrap(int nDim, ...)
 	}
 
 	// Arrays
-	double* input = new double[nDim];                                       // Get the input from the function with unknown number of parameter
+	float* input = new float[nDim];                                       // Get the input from the function with unknown number of parameter
 
 	// Get the input from the function with unknown number of parameters
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 	// Constants that we will use along the function
 	const int N_DIM_MINUS_ONE = nDim - 1;
 	const int N_DIM_PLUS_ONE = nDim + 1;
@@ -439,7 +439,7 @@ int* PerlinNoiseND_Test_Wrap(int nDim, ...)
 	int* unitGridCells = new int[nDim];                                     // Find unit grid cell containing point
 	int* gradientIndices = new int[NUMBER_OF_GRADIENT_INDICES];             // Calculate a set of eight hashed gradient indices
 	int* grad3 = new int[NUMBER_OF_EDGES * nDim];                           // Calculate the gradient helper for noise contributions
-	double* noiseContributions = new double[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
+	float* noiseContributions = new float[NUMBER_OF_GRADIENT_INDICES];    // Calculate noise contributions from each of the eight corners
 
 	for (int i = 0; i < nDim; i++)
 	{
@@ -464,7 +464,7 @@ int* PerlinNoiseND_Test_Wrap(int nDim, ...)
 	return unitGridCells;
 }
 
-int* ClassicPerlinNoise1D_Test_Gradients(double x)
+int* ClassicPerlinNoise1D_Test_Gradients(float x)
 {
 	// Find unit grid cell containing point
 	int xi = FastFloor(x);
@@ -492,9 +492,9 @@ int* PerlinNoiseNDArray_Test_Gradients(int nDim, ...)
 		return 0;
 	}
 
-	double* input = new double[nDim];
+	float* input = new float[nDim];
 
-	VA_LIST_GET(nDim, double, input);
+	VA_LIST_GET(nDim, float, input);
 
 	const int N_DIM_MINUS_ONE = nDim - 1;
 	const int NUMBER_OF_GRADIENT_INDICES = 1 << nDim;
